requirements.txt
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
webdriver-manager>=3.8.0
lxml>=4.9.0
PyYAML>=6.0
plotly>=5.13.0
nbformat>=5.1.3
python-dotenv>=1.0.0
